Split server states in different files
