+++ Assets/Scripts/Server/GameCreationState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9ce2bf89a8d2deb43ae7c402749225b4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
