+++ Assets/Scripts/Server/GameplayState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 802da27c55ee1034dba8bca9ead5ea62
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
